Update spitftbitmap.ino
read16 and read32 should take references to the File object rather than the object itself.  On Yun the File destructor causes the file to be closed when read16/read32 return so future read16/32 calls fail to read any data.  Using a reference prevents the destructor from being called so the file stays open as expected.